# api/docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiMeta": {
        "properties": {
          "page": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "pageSize": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "requestId": {
            "type": "string"
          },
          "total": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "requestId"
        ],
        "type": "object"
      },
      "ArchivedFilter": {
        "enum": [
          "exclude",
          "only",
          "include"
        ],
        "type": "string"
      },
      "BoardResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BoardSnapshotDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "BoardSnapshotDto": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectDto"
          },
          "statuses": {
            "items": {
              "$ref": "#/components/schemas/StatusDto"
            },
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/BoardSummaryDto"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagDto"
            },
            "type": "array"
          },
          "taskTags": {
            "items": {
              "$ref": "#/components/schemas/TaskTagDto"
            },
            "type": "array"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            },
            "type": "array"
          }
        },
        "required": [
          "project",
          "statuses",
          "tasks",
          "tags",
          "taskTags",
          "summary"
        ],
        "type": "object"
      },
      "BoardSummaryDto": {
        "properties": {
          "activeTaskCount": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "archivedTaskCount": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "doneTaskCount": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "activeTaskCount",
          "doneTaskCount",
          "archivedTaskCount"
        ],
        "type": "object"
      },
...
      "CreateTaskRequest": {
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "dueDate": {
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskPriority"
              }
            ]
          },
          "startDate": {
            "type": [
              "string",
              "null"
            ]
          },
          "statusId": {
            "type": "string"
          },
          "tagIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "statusId",
          "title"
        ],
        "type": "object"
      },
...
      "ProjectDto": {
        "properties": {
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "color",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "ReorderTasksRequest": {
        "properties": {
          "destinationStatusId": {
            "type": "string"
          },
          "movedTaskId": {
            "type": "string"
          },
          "orderedTaskIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "sourceStatusId": {
            "type": "string"
          }
        },
        "required": [
          "movedTaskId",
          "sourceStatusId",
          "destinationStatusId",
          "orderedTaskIds"
        ],
        "type": "object"
      },
      "SortOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
      "StatusDto": {
        "properties": {
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isDone": {
            "type": "boolean"
          },
          "isHidden": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "color",
          "sortOrder",
          "isDone",
          "isHidden",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "TagDto": {
        "properties": {
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "color",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "TaskDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "dueDate": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "position": {
            "format": "int64",
            "type": "integer"
          },
          "priority": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskPriority"
              }
            ]
          },
          "projectId": {
            "type": "string"
          },
          "startDate": {
            "format": "date",
            "type": "string"
          },
          "statusId": {
            "type": "string"
          },
          "tagIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "statusId",
          "title",
          "position",
          "startDate",
          "dueDate",
          "completedAt",
          "archivedAt",
          "tagIds",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "TaskListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "TaskPriority": {
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "type": "string"
      },
      "TaskResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "TaskSortBy": {
        "enum": [
          "updatedAt",
          "dueDate",
          "createdAt",
          "position"
        ],
        "type": "string"
      },
      "TaskTagDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "tagId": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          }
        },
        "required": [
          "projectId",
          "taskId",
          "tagId",
          "createdAt"
        ],
        "type": "object"
      },
...
      "UpdateTaskRequest": {
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "dueDate": {
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskPriority"
              }
            ]
          },
          "startDate": {
            "type": [
              "string",
              "null"
            ]
          },
          "statusId": {
            "type": [
              "string",
              "null"
            ]
          },
          "tagIds": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local kanban task management API",
    "license": {
      "name": ""
    },
    "title": "Plan API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{projectId}/board": {
      "get": {
        "operationId": "getBoardSnapshot",
        "parameters": [
          {
            "description": "Project id",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeArchived",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/api/v1/projects/{projectId}/tasks": {
      "get": {
        "operationId": "listProjectTasks",
        "parameters": [
          {
            "description": "Project id",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "statusId",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/TaskPriority"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "tagId",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/ArchivedFilter"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/TaskSortBy"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/SortOrder"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "post": {
        "operationId": "createProjectTask",
        "parameters": [
          {
            "description": "Project id",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/projects/{projectId}/tasks/reorder": {
      "post": {
        "operationId": "reorderProjectTasks",
        "parameters": [
          {
            "description": "Project id",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTasksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/api/v1/tasks/{taskId}": {
      "delete": {
        "operationId": "deleteTask",
        "parameters": [
          {
            "description": "Task id",
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "get": {
        "operationId": "getTask",
        "parameters": [
          {
            "description": "Task id",
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "operationId": "updateTask",
        "parameters": [
          {
            "description": "Task id",
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/tasks/{taskId}/archive": {
      "post": {
        "operationId": "archiveTask",
        "parameters": [
          {
            "description": "Task id",
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/api/v1/tasks/{taskId}/restore": {
      "post": {
        "operationId": "restoreTask",
        "parameters": [
          {
            "description": "Task id",
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://localhost:3001/api/v1"
    }
  ],
...
}